Load checkpoints in memoizer code, rather than DFK
This is first of a set of PRs to move checkpoint handling code out of
the DFK in eventually-pluggable memoizers.
